requirements.txt
# --- Core Application & UI Framework ---
streamlit==1.33.0
streamlit-option-menu==0.3.12

# --- Database Connector ---
psycopg2-binary==2.9.9

# --- Deep Learning & NLP ---
torch==2.2.2
nltk==3.8.1
gensim==4.3.2
google-generativeai==0.5.0
contractions==0.1.73
emoji==2.11.0

# --- Data Handling & Plotting ---
numpy==1.26.4
pandas==2.2.1
plotly==5.20.0

# --- Utilities & Security ---
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.2